Wrong warning fixes correct in ccUTF8.cpp.

+++ cocos/2d/ccUTF8.h
@@ -80,9 +80,9 @@ CC_DLL unsigned short* cc_utf8_to_utf16(const char* str_old, int length = -1, in
  **/
 CC_DLL char *
 cc_utf16_to_utf8 (const unsigned short  *str,
-                  long             len,
-                  long            *items_read,
-                  long            *items_written);
+                  ssize_t             len,
+                  ssize_t            *items_read,
+                  ssize_t            *items_written);
 
 NS_CC_END
